Moving newReports into report package
